refactor(geo): relocate Yahoo geocoder components to `lib.geo` for improved modularity and consistency
