+++ lib/geo/valueobject/yahoo.py
